+++ packages/general/src/crypto/index.ts
@@ -9,6 +9,7 @@ export * from "./CryptoConstants.js";
 export * from "./CryptoError.js";
 export * from "./Key.js";
 export * from "./MockCrypto.js";
+export * from "./NodeJsStyleCrypto.js";
 export * from "./Spake2p.js";
 export * from "./StandardCrypto.js";
 export * from "./WebCrypto.js";
