+++ packages/nodejs/src/crypto/NodeJsCrypto.ts
@@ -4,183 +4,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {
-    Bytes,
-    CRYPTO_AUTH_TAG_LENGTH,
-    CRYPTO_EC_CURVE,
-    CRYPTO_EC_KEY_BYTES,
-    CRYPTO_ENCRYPT_ALGORITHM,
-    CRYPTO_HASH_ALGORITHM,
-    CRYPTO_SYMMETRIC_KEY_LENGTH,
-    Crypto,
-    CryptoDecryptError,
-    CryptoDsaEncoding,
-    CryptoVerifyError,
-    PrivateKey,
-    PublicKey,
-    asError,
-} from "#general";
+import { NodeJsStyleCrypto } from "#general";
 import * as crypto from "node:crypto";
 
-export class NodeJsCrypto extends Crypto {
-    implementationName = "Node.js";
-
-    encrypt(key: Bytes, data: Bytes, nonce: Bytes, aad?: Bytes): Bytes {
-        const cipher = crypto.createCipheriv(CRYPTO_ENCRYPT_ALGORITHM, Bytes.of(key), Bytes.of(nonce), {
-            authTagLength: CRYPTO_AUTH_TAG_LENGTH,
-        });
-        if (aad !== undefined) {
-            cipher.setAAD(Bytes.of(aad), { plaintextLength: data.byteLength });
-        }
-        const encrypted = cipher.update(Bytes.of(data));
-        cipher.final();
-        return Bytes.concat(Bytes.of(encrypted), Bytes.of(cipher.getAuthTag()));
-    }
-
-    decrypt(key: Bytes, encrypted: Bytes, nonce: Bytes, aad?: Bytes): Bytes {
-        const cipher = crypto.createDecipheriv(CRYPTO_ENCRYPT_ALGORITHM, Bytes.of(key), Bytes.of(nonce), {
-            authTagLength: CRYPTO_AUTH_TAG_LENGTH,
-        });
-        const data = Bytes.of(encrypted);
-        const plaintextLength = data.length - CRYPTO_AUTH_TAG_LENGTH;
-        if (aad !== undefined) {
-            cipher.setAAD(Bytes.of(aad), { plaintextLength });
-        }
-        cipher.setAuthTag(data.slice(plaintextLength));
-        const result = cipher.update(data.slice(0, plaintextLength));
-        try {
-            cipher.final();
-        } catch (e) {
-            throw new CryptoDecryptError(`${CRYPTO_ENCRYPT_ALGORITHM} decryption failed: ${asError(e).message}`);
-        }
-        return Bytes.of(result);
-    }
-
-    randomBytes(length: number): Bytes {
-        return Bytes.of(crypto.randomBytes(length));
-    }
-
-    ecdhGeneratePublicKey(): { publicKey: Bytes; ecdh: any } {
-        const ecdh = crypto.createECDH(CRYPTO_EC_CURVE);
-        ecdh.generateKeys();
-        return { publicKey: Bytes.of(ecdh.getPublicKey()), ecdh: ecdh };
-    }
-
-    ecdhGeneratePublicKeyAndSecret(peerPublicKey: Bytes): {
-        publicKey: Bytes;
-        sharedSecret: Bytes;
-    } {
-        const ecdh = crypto.createECDH(CRYPTO_EC_CURVE);
-        ecdh.generateKeys();
-        return {
-            publicKey: Bytes.of(ecdh.getPublicKey()),
-            sharedSecret: Bytes.of(ecdh.computeSecret(Bytes.of(peerPublicKey))),
-        };
-    }
-
-    computeSha256(data: Bytes | Bytes[]): Bytes {
-        const hasher = crypto.createHash(CRYPTO_HASH_ALGORITHM);
-        if (Array.isArray(data)) {
-            data.forEach(chunk => hasher.update(Bytes.of(chunk)));
-        } else {
-            hasher.update(Bytes.of(data));
-        }
-        return Bytes.of(hasher.digest());
-    }
-
-    createPbkdf2Key(secret: Bytes, salt: Bytes, iteration: number, keyLength: number): Promise<Bytes> {
-        return new Promise<Bytes>((resolver, rejecter) => {
-            crypto.pbkdf2(
-                Bytes.of(secret),
-                Bytes.of(salt),
-                iteration,
-                keyLength,
-                CRYPTO_HASH_ALGORITHM,
-                (error, key) => {
-                    if (error !== null) rejecter(error);
-                    resolver(Bytes.of(key));
-                },
-            );
-        });
-    }
-
-    createHkdfKey(
-        secret: Bytes,
-        salt: Bytes,
-        info: Bytes,
-        length: number = CRYPTO_SYMMETRIC_KEY_LENGTH,
-    ): Promise<Bytes> {
-        return new Promise<Bytes>((resolver, rejecter) => {
-            crypto.hkdf(
-                CRYPTO_HASH_ALGORITHM,
-                Bytes.of(secret),
-                Bytes.of(salt),
-                Bytes.of(info),
-                length,
-                (error, key) => {
-                    if (error !== null) rejecter(error);
-                    resolver(Bytes.of(key));
-                },
-            );
-        });
-    }
-
-    signHmac(key: Bytes, data: Bytes): Bytes {
-        const hmac = crypto.createHmac(CRYPTO_HASH_ALGORITHM, Bytes.of(key));
-        hmac.update(Bytes.of(data));
-        return Bytes.of(hmac.digest());
-    }
-
-    signEcdsa(privateKey: JsonWebKey, data: Bytes | Bytes[], dsaEncoding: CryptoDsaEncoding = "ieee-p1363"): Bytes {
-        const signer = crypto.createSign(CRYPTO_HASH_ALGORITHM);
-        if (Array.isArray(data)) {
-            data.forEach(chunk => signer.update(Bytes.of(chunk)));
-        } else {
-            signer.update(Bytes.of(data));
-        }
-        return Bytes.of(
-            signer.sign({
-                key: privateKey as any,
-                format: "jwk",
-                type: "pkcs8",
-                dsaEncoding,
-            }),
-        );
-    }
-
-    verifyEcdsa(publicKey: JsonWebKey, data: Bytes, signature: Bytes, dsaEncoding: CryptoDsaEncoding = "ieee-p1363") {
-        const verifier = crypto.createVerify(CRYPTO_HASH_ALGORITHM);
-        verifier.update(Bytes.of(data));
-        const success = verifier.verify(
-            {
-                key: publicKey as any,
-                format: "jwk",
-                type: "spki",
-                dsaEncoding,
-            },
-            Bytes.of(signature),
-        );
-        if (!success) throw new CryptoVerifyError("Signature verification failed");
-    }
-
-    createKeyPair() {
-        // Note that we this key may be used for DH or DSA but we use an ECDH to generate
-        const ecdh = crypto.createECDH(CRYPTO_EC_CURVE);
-        ecdh.generateKeys();
-
-        // The key exported from Node doesn't include most-significant bytes that are 0.  This doesn't affect how we
-        // currently use keys but it's a little weird so 0 pad to avoid future confusion
-        const privateKey = new Uint8Array(CRYPTO_EC_KEY_BYTES);
-        const nodePrivateKey = ecdh.getPrivateKey();
-        privateKey.set(nodePrivateKey, CRYPTO_EC_KEY_BYTES - nodePrivateKey.length);
-
-        return PrivateKey(privateKey, { publicKey: Bytes.of(ecdh.getPublicKey()) });
-    }
-
-    generateDhSecret(key: PrivateKey, peerKey: PublicKey): Bytes {
-        const ecdh = crypto.createECDH(CRYPTO_EC_CURVE);
-        ecdh.setPrivateKey(Bytes.of(key.privateBits));
-
-        return Bytes.of(ecdh.computeSecret(Bytes.of(peerKey.publicBits)));
+/**
+ * Node.js-based crypto implementation.
+ */
+export class NodeJsCrypto extends NodeJsStyleCrypto {
+    constructor() {
+        super(crypto);
     }
 }
